requirements.txt
ansible-core
requests